rename the `get` function to `entries`
This is because `get` is usually the name of a method that fetches
a single value from a collection according to the given query.
In contrast, Minissg's `get` functions must return a module, which can
be considered as an entire collection.
For such purpose, `entries` is usually used in JavaScript community.

BREAKING CHANGE: The `get` function has been renamed to `entries`.
  Users must rename their `get` functions by your hands.
  No migration support is provided.