add "next" branch for prereleases
It is strongly recommended to do prerelease before releasing a stable
release because we cannot cancel publishing a package n npm.js; once
a package is uploaded, its version number is occupied forever even if
it is unpublished.  Making a stable version number unavailable due to
failure of uploading or packaging is quite unsatisfactory.  Prerelease
allows us to make sure that a package can be successfully created and
uploaded before stable release.