fix: prepend `site.projectRoot` to the resolved id of Head module
I don't know why but Vite's html plugin requires that HTML file must be
in `config.root` and identified with its absolute path, or Rollup causes
an error complaining that an illegal path is passed to this.emitFile.
The cause of this is the following code in vite/src/node/plugins/html.ts:

    const shortEmitName = normalizePath(path.relative(config.root, id))
    this.emitFile({
      type: 'asset',
      fileName: shortEmitName,
      source: result,
    })

To avoid this, we make the id of Head virtual modules a fake absolute
path in `config.root` by prepending `site.projectRoot` to its resolved
(`\0`-included) id.