put just a single <script> or <link> in <head> of generated html files
Putting the list of all the module reference in HTML results in an
unoptimized chunks; in particular, some empty chunks may remain.
Because HTML files are dealt with by Vite as non-treeshakable modules,
its every child module has some side effect even if it is actually empty.
Concequently, if a child module of HTML is imported from multiple modules,
the module forms an individual chunk that are referenced from those
modules evan if it is empty.

This change fixes this issue since this avoids any module that is not
shared among any pair of HTML files.