remove Keep
If I remember correctly, Keep was introduced so that we can prevent
Rollup from deduplicating Head modules.  We expect that Keep is not
needed anymore since we now have for each html file a unique module
that aggregates imports of dependencies.

Instead of Keep, Lib is used as a dummy entry module at the client-side
run when any entry does not generate any asset.  It is safe because
Lib is never used in the client-side run, but just a piece of hack.