express immutability of Module and its related types by adding `readonly`
We didn't add `readonly` to Module's properties but we add it to an array
type in Module because, in TypeScript, readonly object can be seen as
identical to non-readonly object, whereas readonly array type is distinct
from non-readonly array type (readonly array doesn't have `push` and `pop`,
for example).  So, adding `readonly` carefully is needed for arrays but
not needed for object properties.

We add `readonly` to EntriesArg so that the user can notice that the
argument of `entries` function is not intended to be modified in the
body of `entries`.  We don't specify what will happen when the argument
is destructively updated.