redesign and rewite @minissg/page at all
The new page library is designed in a functional style: imperative
tree construction methods such as bind and mount are removed.
After constructing a Page object, any member of it are never changed.
Pages are searched for by tree traversing.
No cache mechanism is included so that Vite's HMR works as expected.