imprement memoization of promises created by Page
This memoization is not only for performance but for React suspense,
which requires the management of identity of promises.  The memoized
promises are thrown away when either a new Page is created and therefore
the page tree has been possibly changed or the rendering of a new page
is started, so that Minissg can compute the assets loaded for each page
based on the side effect of dynamic import.