stop forgetting memoized data at Page.constructor.
This is not a appropriate place for clearing memoization.
Memo must be persistent during a render session.
It must be assumed that the structure of page tree is also persistent
and static at least during a session.  In this sense, creating a Page
is just a delayed computation of the static page tree.