rename the `entries` function to `main`
From the experience of using the names `get` and `entries` for lazy
construction of module tree, we've found that they can be easily
confused with standard methods of the same name in both sense of
readability and consistency.  We've concluded that the name must be
chosen from ones that are short but not (or rarely) used in the
standard or well-known libraries.  Consequently, we choose `main`.
It is short and doesn't look like a library function but a entry
point of a program, which fits with the purpose of the function in
Minissg.

BREAKING CHANGE: The `entries` function has been renamed to `main`.
  Users must rename their `entries` functions by your hands.